chore: separate model field, db connection info
